event listeners as services
each wollmux event listener is a service
