+++ src/main/java/de/muenchen/allg/itd51/wollmux/event/CheckInstallationListener.java
@@ -4,11 +4,16 @@ import com.google.common.eventbus.Subscribe;
 
 import de.muenchen.allg.itd51.wollmux.event.handlers.OnCheckInstallation;
 
-public class CheckInstallation
+public class CheckInstallationListener implements WollMuxEventListener
 {
+  public CheckInstallationListener()
+  {
+  }
+
   @Subscribe
   public void onCheckInstallation(OnCheckInstallation event)
   {
+    WollMuxEventHandler.getInstance().unregisterListener(this);
     event.process();
   }
 }
