+++ src/main/java/de/muenchen/allg/itd51/wollmux/event/InitEventListener.java
@@ -4,11 +4,16 @@ import com.google.common.eventbus.Subscribe;
 
 import de.muenchen.allg.itd51.wollmux.event.handlers.OnInitialize;
 
-public class InitEventListener
+public class InitEventListener implements WollMuxEventListener
 {
+  public InitEventListener()
+  {
+  }
+
   @Subscribe
   public void onInitialize(OnInitialize event)
   {
+    WollMuxEventHandler.getInstance().unregisterListener(this);
     event.process();
   }
 }
