+++ src/main/java/de/muenchen/allg/itd51/wollmux/event/handlers/OnCheckInstallation.java
@@ -18,7 +18,6 @@ import de.muenchen.allg.afid.UNO;
 import de.muenchen.allg.itd51.wollmux.WollMuxFehlerException;
 import de.muenchen.allg.itd51.wollmux.core.util.L;
 import de.muenchen.allg.itd51.wollmux.dialog.InfoDialog;
-import de.muenchen.allg.itd51.wollmux.event.WollMuxEventHandler;
 
 /**
  * Erzeugt ein neues WollMux-Event, in dem geprüft wird, ob der WollMux korrekt
@@ -36,7 +35,6 @@ public class OnCheckInstallation extends BasicEvent
   @Override
   protected void doit() throws WollMuxFehlerException
   {
-    WollMuxEventHandler.getInstance().unregisterCheckInstallationListener();
     // Standardwerte für den Warndialog:
     boolean showdialog = true;
     String title = L.m("Mehrfachinstallation des WollMux");
