+++ src/main/resources/META-INF/services/de.muenchen.allg.itd51.wollmux.event.WollMuxEventListener
@@ -1,0 +1,4 @@
+de.muenchen.allg.itd51.wollmux.event.WollMuxEventListenerImpl
+de.muenchen.allg.itd51.wollmux.event.InitEventListener
+de.muenchen.allg.itd51.wollmux.event.CheckInstallationListener
+de.muenchen.allg.itd51.wollmux.slv.events.ContentBasedDirectiveEventListener
